+++ Reader/MongoDBODM/DeltaConfigurableReader.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Pim\Bundle\MagentoConnectorBundle\Reader\MongoDBODM;
+
+use Pim\Bundle\BaseConnectorBundle\Reader\Doctrine\ODMProductReader;
+
+/**
+ * @author Romain Monceau <romain@akeneo.com>
+ */
+class DeltaConfigurableReader extends ODMProductReader
+{
+}