+++ Reader/MongoDBODM/DeltaProductReader.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Pim\Bundle\MagentoConnectorBundle\Reader\MongoDBODM;
+
+use Pim\Bundle\BaseConnectorBundle\Reader\Doctrine\ODMProductReader;
+
+/**
+ * Delta product reader
+ *
+ * @author Romain Monceau <romain@akeneo.com>
+ */
+class DeltaProductReader extends ODMProductReader
+{
+}