+++ CHANGELOG.md
@@ -16,6 +16,10 @@
  - CategoryRepository::findOrderCategories takes a CategoryInterface
  - pim_base_connector.reader.doctrine.obsoleteproduct, pim_base_connector.reader.doctrine.obsoleteproduct.class, pim_magento_connector.reader.doctrine.product has been removed from readers.yml because they are unused
 
+ - Rename parameter `pim_magento_connector.reader.orm.delta_product.class` by `pim_magento_connector.reader.delta_product.class`
+ - Rename parameter `pim_magento_connector.reader.orm.delta_product_association.class` by `pim_magento_connector.reader.delta_product_association.class`
+ - Rename parameter `pim_magento_connector.reader.orm.delta_product.class` by `pim_magento_connector.reader.delta_product.class`
+
 # 1.1.20 (2015-02-20)
 ## Bug fixes
  - Update product and category url keys in admin store view when multiple magento store views are used.
@@ -36,7 +40,7 @@
 # 1.1.17 (2015-02-16)
 ## Bug fixes
  - Prevent error during product creation if sku is a number.
- 
+
 # 1.1.16 (2015-02-13)
 ## Bug fixes
  - Url_key are now correctly handled when multiple magento store views are used.